harvestmind-platform-main/python_version/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.0
Flask-Login==0.6.2
Werkzeug==2.3.7
pyjwt==2.8.0
supabase==2.0.2
requests==2.31.0
python-dateutil==2.8.2 